Switch EstimateSolution and GetAvarage return types from int to double

+++ include/EvoAlg.h
@@ -36,7 +36,7 @@ public:
   void Evolve();
   Solution *GetBest();
   Solution *GetWorst();
-  int GetAvarage();
+  double GetAvarage();
 
 protected:
   Solution *Select();
